+++ lvis-api/apps/warehouse/src/seriv-item/entities/seriv-item.entity.ts
@@ -14,7 +14,7 @@ export class SERIVItem {
   @Field(() => String)
   item_id: string;
 
-  @Field(() => Int)
+  @Field(() => Float)
   quantity: number;
 
   @Field(() => Float)
